[OWA-932] Revert back to webpack, transpileOnly
Parcel is super fast, but I was getting bogged down by issue after issue. So gave up and went back to webpack. Sadly, webpack is as slow as tracle if it needs to statically analyze each file on build.  So it’s transpileOnly now, meaning you’ll only get type errors in your IDE for now.